guarantee that dependencies dont get re-exported
Here's how ngimport is executed:

1. ngimport's bootstrap is run, which exports an instance of each
provider
2. When the host application's bootstrap is run, ngimport's run and
config blocks are run again (!)
3. This causes stateful services ($httpBackend, $rootScope) to be
re-exported!

This causes a subtle bug where an ngimported dependency will be one of
two possible instances, depending on when it is used (before the host
application is bootstrapped vs. after).

This commit ensures that the 1st exported version is the only one that
is exposed.